+++ apps/api/services/connect_service.py
@@ -3,21 +3,19 @@ from apps.core.models import Connect
 from .branch_service import verification_owner_branch_service
 
 
-def get_connect_by_id_service(user, connect_id):
+def get_connect_by_id_service(user: object, connect_id: int) -> object | None:
     """
-    Функция возвращает connect по connect_id если пользователь является owner в company, иначе вернет False
+    Функция возвращает connect по connect_id если пользователь является owner в company, иначе вернет None
     :param user: Текущий пользователь
     :param connect_id: id соединения
-    :return: connect или False
+    :return: optional: connect
     """
-    connect = Connect.objects.get(pk=connect_id)
-    if verification_owner_branch_service(branch_id=connect.branch_id, user=user):
-        return connect
-    else:
-        return False
+    connect: object = Connect.objects.get(pk=connect_id)
+    result: object | None = (None, connect)[verification_owner_branch_service(branch_id=connect.branch_id, user=user)]
+    return result
 
 
-def get_connect_by_id_branch_and_type(user, branch_id, connect_type):
+def get_connect_by_id_branch_and_type(user: object, branch_id: int, connect_type: int) -> object:
     """
     Функция возвращает connect с указанным типом
     :param user: Текущий пользователь
@@ -25,20 +23,24 @@ def get_connect_by_id_branch_and_type(user, branch_id, connect_type):
     :param connect_type: тип connect
     :return: connect list
     """
-    connect = Connect.objects.filter(branch_id=branch_id, type=connect_type).last()
+    connect: object = Connect.objects.filter(branch_id=branch_id, type=connect_type).last()
     return connect
 
 
-def create_connect_by_branch_id_service(user, branch_id, connect_type, key):
+def create_connect_by_branch_id_service(user: object, branch_id: int, connect_type: int, key: int) -> object | None:
     """
-    Создает новый connect если пользователь является owner в company, иначе вернет False
+    Создает новый connect если пользователь является owner в company, иначе вернет None
     :param user: Текущий пользователь
     :param branch_id: id branch
     :param connect_type: Тип платформы
     :param key: Ключ для соединения с площадкой
-    :return: connect или False
+    :return: optional: connect
     """
-    if verification_owner_branch_service(user=user, branch_id=branch_id):
-        return Connect.objects.create(branch_id=branch_id, type=connect_type, key=key)
-    else:
-        return False
+    result: object | None = (
+        None,
+        Connect.objects.create(
+            branch_id=branch_id,
+            type=connect_type, key=key
+        )
+    )[verification_owner_branch_service(user=user, branch_id=branch_id)]
+    return result