+++ apps/api/services/rate_info_service.py
@@ -1,7 +1,7 @@
 from apps.core.models import RateInfo
 
 
-def get_rate_info_service():
+def get_rate_info_service() -> object:
     """
     Функция возвращает последний rate info
     :return: rate info